Dockerfile runs IOC basler-acA1300-75gm from aravisGigE image
The previous code of the Dockerfile in this repository has been split into two distinct Dockerfiles: one for the aravisGigE EPICS module and another for the camera IOC (basler-acA1300-75gm-epics-ioc).

This repository contains the Dockerfile for running the camera IOC. The repository with the Dockerfile for creating the aravisGigE image is named docker-aravisGigE-epics-module.